Split EnvoyUtil into Resources.Common and Resources.LDS
